命令文件修改

+++ docs/install.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:~/bin
+PATH=$PATH:/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:~/bin:$(cd `dirname $0`; pwd)
 export PATH
 LANG=en_US.UTF-8
 
@@ -9,12 +9,30 @@ if [[ ! -f "/usr/bin/unzip" ]];then
 	yum install unzip -y
 fi
 TYPE="$1"
+
+module="$2"
+# 判断是否包含jdk
+if [[ "$module" = "jdk" ]]; then
+  # 判断是否存在文件
+  if [[ ! -f "jdk-8u251-linux-x64.tar.gz" ]]; then
+   wget -O jdk-8u251-linux-x64.tar.gz https://jpom-releases.oss-cn-hangzhou.aliyuncs.com/jdk-8u251-linux-x64.tar.gz
+  fi
+  #
+  tar -zxvf jdk-8u251-linux-x64.tar.gz
+  #
+  PATH=$PATH:$(cd `dirname $0`; pwd)/jdk1.8.0_251/bin
+  export PATH
+fi
+
 # 判断
 if [[ -z "${TYPE}" ]] ; then
     TYPE="Server";
 fi
-# 下载
-wget -O ${TYPE}.zip https://1232788122276831.cn-beijing.fc.aliyuncs.com/2016-08-15/proxy/jpom/jpom-releases/?type=${TYPE}
+# 判断是否在文件
+if [[ ! -f "${TYPE}.zip" ]]; then
+  # 下载
+  wget -O ${TYPE}.zip https://1232788122276831.cn-beijing.fc.aliyuncs.com/2016-08-15/proxy/jpom/jpom-releases/?type=${TYPE}
+fi
 # 解压
 unzip -o ${TYPE}.zip
 # 删除安装包
